requirements.txt
APScheduler==3.8.1
arabic-reshaper==2.1.3
asgiref==3.2.10
awscli==1.19.112
boto3==1.17.92
botocore==1.20.112
CacheControl==0.12.10
cachetools==4.2.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==4.0.3
cryptography==3.4.7
decorator==5.0.9
Django==3.1.1
django-cors-headers==3.5.0
django-environ==0.7.0
django-jsonfield==1.4.1
djangorestframework==3.11.1
docutils==0.15.2
drf-yasg==1.20.0
firebase-admin==5.2.0
future==0.18.2
geocoder==1.38.1
google-api-core==2.6.1
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.3.4
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.3.1
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
html5lib==1.1
httpagentparser==1.9.1
httplib2==0.20.4
idna==2.10
importlib-metadata==4.3.1
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
json2xml==3.16.0
MarkupSafe==1.1.1
msgpack==1.0.3
numpy==1.19.5
openapi-codec==1.3.2
packaging==20.9
Pillow==8.3.1
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-http-client==3.3.2
pytz==2021.1
PyYAML==5.3.1
ratelim==0.1.6
reportlab==3.6.6
requests==2.25.1
rsa==4.5
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
sendgrid==6.7.0
simplejson==3.17.2
six==1.15.0
soft-delete==0.2.2
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
stripe==2.64.0
typing-extensions==3.10.0.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
xhtml2pdf==0.2.5
xmltodict==0.12.0
zipp==3.4.1
